pyproject.toml
[project]
name = "pdf-toc-gen"
version = "0.1.0"
description = "CLI tool to add TOC bookmarks to PDFs"
authors = [
    {name = "Kai Erik Niermann",email = "kai.niermann@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf>=1.24.0",
    "typer>=0.12.0",
    "ocrmypdf>=16.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
pdftoc = "pdftoc.cli:run"

[tool.poetry]
packages = [{include = "pdftoc", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
